Remove underscore in _build* util functions. Move load_config outside agent subpackage.
