+++ froggy/tests/test_agents.py
@@ -1,11 +1,8 @@
-from unittest.mock import patch
-
 import pytest
 
 from froggy.agents.utils import (
     HistoryTracker,
-    _build_history_prompt,
-    load_config,
+    build_history_prompt,
     trim_prompt_messages,
 )
 
@@ -214,71 +211,6 @@ def test_history_tracker():
     assert ht.json() == {}
 
 
-def test_load_config():
-    import atexit
-    import tempfile
-    from pathlib import Path
-
-    import yaml
-
-    # do the test in a tmp folder
-    tempdir = tempfile.TemporaryDirectory(prefix="TestLoadConfig-")
-    working_dir = Path(tempdir.name)
-    config_file = working_dir / "config.yaml"
-    atexit.register(tempdir.cleanup)  # Make sure to cleanup that folder once done.
-
-    config_contents = {}
-    config_contents["zero_shot"] = {
-        "random_seed": 42,
-        "max_steps": 100,
-        "llm_name": "gpt2",
-        "llm_temperature": [0.5],
-    }
-    config_contents["cot"] = {
-        "random_seed": 43,
-        "max_steps": 50,
-        "cot_style": "standard",
-        "llm_name": "gpt20",
-        "llm_temperature": [0.3, 0.5],
-    }
-
-    # write the config file into yaml
-    with open(config_file, "w") as f:
-        yaml.dump(config_contents, f)
-
-    # now test
-    with patch(
-        "sys.argv",
-        [
-            "config_file",
-            str(config_file),
-            "--agent",
-            "zero_shot",
-            "-p",
-            "zero_shot.random_seed=123",
-            "cot.llm_temperature=[0.8, 0.8]",
-            "-v",
-            "--debug",
-        ],
-    ):
-
-        _config, _args = load_config()
-    assert _args.agent == "zero_shot"
-    assert "zero_shot" in _config.keys()
-    assert "cot" in _config.keys()
-    assert _config["zero_shot"]["random_seed"] == 123
-    assert _config["zero_shot"]["max_steps"] == 100
-    assert _config["zero_shot"]["llm_name"] == "gpt2"
-    assert _config["zero_shot"]["llm_temperature"] == [0.5]
-    assert _config["cot"]["random_seed"] == 43
-    assert _config["cot"]["max_steps"] == 50
-    assert _config["cot"]["cot_style"] == "standard"
-    assert _config["cot"]["llm_name"] == "gpt20"
-    assert _config["cot"]["llm_temperature"] == [0.8, 0.8]
-    assert _args.debug is True
-    assert _args.verbose is True
-
-
 def test_build_history_prompt():
     import json
 
@@ -287,13 +219,13 @@ def test_build_history_prompt():
     # test with empty history
     ht = HistoryTracker(history_steps=3)
     # use_conversational_prompt is False
-    messages = _build_history_prompt(ht, use_conversational_prompt=False)
+    messages = build_history_prompt(ht, use_conversational_prompt=False)
     expected = [
         {"role": "user", "content": "No history of command and terminal outputs."}
     ]
     assert messages == expected
     # use_conversational_prompt is True
-    messages = _build_history_prompt(ht, use_conversational_prompt=True)
+    messages = build_history_prompt(ht, use_conversational_prompt=True)
     expected = [
         {"role": "user", "content": "No history of command and terminal outputs."}
     ]
@@ -318,7 +250,7 @@ def test_build_history_prompt():
 
     # use_conversational_prompt is False
     # reset_prompt_history_after_rewrite is False
-    messages = _build_history_prompt(
+    messages = build_history_prompt(
         ht, use_conversational_prompt=False, reset_prompt_history_after_rewrite=False
     )
     expected = [f"History of command and terminal outputs (the last 3 steps):"]
@@ -332,7 +264,7 @@ def test_build_history_prompt():
     assert messages == expected
 
     # reset_prompt_history_after_rewrite is True
-    messages = _build_history_prompt(
+    messages = build_history_prompt(
         ht, use_conversational_prompt=False, reset_prompt_history_after_rewrite=True
     )
     expected = [f"History of command and terminal outputs (the last 2 steps):"]
@@ -346,7 +278,7 @@ def test_build_history_prompt():
 
     # use_conversational_prompt is True
     # reset_prompt_history_after_rewrite is False
-    messages = _build_history_prompt(
+    messages = build_history_prompt(
         ht, use_conversational_prompt=True, reset_prompt_history_after_rewrite=False
     )
     expected = [
@@ -366,7 +298,7 @@ def test_build_history_prompt():
     expected += history_messages
     assert messages == expected
     # reset_prompt_history_after_rewrite is True
-    messages = _build_history_prompt(
+    messages = build_history_prompt(
         ht, use_conversational_prompt=True, reset_prompt_history_after_rewrite=True
     )
     expected = [
